move Prelude and PlatformPackagePath to io
