+++ cli/.js/src/main/scala/aqua/script/ScriptOpts.scala
@@ -6,6 +6,7 @@ import aqua.backend.Generated
 import aqua.backend.air.{AirBackend, AirGen, FuncAirGen}
 import aqua.builder.ArgumentGetter
 import aqua.compiler.AquaCompiler
+import aqua.io.Prelude
 import aqua.ipfs.js.IpfsApi
 import aqua.keypair.KeyPairShow.show
 import aqua.model.transform.{Transform, TransformConfig}
@@ -15,7 +16,17 @@ import aqua.raw.ops.{Call, CallArrowRawTag}
 import aqua.raw.value.{LiteralRaw, ValueRaw, VarRaw}
 import aqua.res.{AquaRes, FuncRes}
 import aqua.run.RunOpts.logger
-import aqua.run.{CliFunc, GeneralOptions, GeneralOpts, RunCommand, RunConfig, RunOpts}
+import aqua.run.{
+  CliFunc,
+  FuncCompiler,
+  GeneralOptions,
+  GeneralOpts,
+  PackagePath,
+  RelativePath,
+  RunCommand,
+  RunConfig,
+  RunOpts
+}
 import aqua.types.{ArrowType, LiteralType, NilType, ScalarType}
 import cats.data.*
 import cats.data.Validated.{invalid, invalidNec, valid, validNec, validNel}
